[#24] Add new RequestHandler middleware
